API/requirements.txt
django
djangorestframework
django_cors_headers
whitenoise